fix(core): restore model line budget
